requirements.txt
python-telegram-bot==20.6
openai>=1.30.0
Flask==3.0.3
python-dotenv==1.0.1